DevApp1/requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.2.post1
cffi==1.17.1
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
db==0.1.1
db-sqlite3==0.0.1
ecdsa==0.19.0
fastapi==0.115.4
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
MarkupSafe==3.0.2
pillow==11.0.0
pwinput==1.0.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
PyYAML==6.0.2
reportlab==4.2.5
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tabulate==0.9.0
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
